requirements.txt
flask==2.3.3
streamlit==1.28.1
supabase==2.0.2
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.0
openpyxl==3.1.2
xlrd==2.0.1
nltk==3.8.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
torch==2.0.1
google-generativeai==0.3.2 